requirements.txt
absl-py==0.9.0
asgiref==3.2.9
astor==0.8.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-cors-headers==3.4.0
django-import-export==2.2.0
django-restframework-apiview==1.4.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.1
jdcal==1.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupPy==1.14
numpy==1.18.5
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
Pillow-PIL==0.1.dev0
protobuf==3.12.2
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
